# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Tarefas",
    "version": "1.0.0",
    "description": "Uma API simples para gerenciar tarefas"
  },
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Lista todas as tarefas",
        "responses": {
          "200": {
            "description": "Retorna a lista de tarefas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tarefa"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria uma nova tarefa",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tarefa"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarefa criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarefa"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "summary": "Busca uma tarefa pelo ID",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna a tarefa encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarefa"
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada"
          }
        }
      },
      "put": {
        "summary": "Atualiza uma tarefa existente",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tarefa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarefa atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tarefa"
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada"
          }
        }
      },
      "delete": {
        "summary": "Remove uma tarefa pelo ID",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarefa removida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Tarefa": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "dt_inicio": {
            "type": "string"
          },
          "comentarios": {
            "type": "string"
          },
          "dificuldade": {
            "type": "string"
          }
        }
      }
    }
  }
}